Rename 'github_workflow.py' model file to 'github_action.py'
